requirements.txt
pytz
python-telegram-bot
pandas
matplotlib
numpy
requests